+++ src/main/java/com/stratio/qa/specs/CCTSpec.java
@@ -955,6 +955,7 @@ public class CCTSpec extends BaseGSpec {
     /**
      * Install service
      * @param service   service name
+     * @param folder    folder where service are going to be installed
      * @param model     service model
      * @param version   service version
      * @param name      service instance name
@@ -962,8 +963,8 @@ public class CCTSpec extends BaseGSpec {
      * @param jsonFile  marathon json to deploy
      * @throws Exception
      */
-    @Given("^I install service '(.+?)' with model '(.+?)' and version '(.+?)' and instance name '(.+?)' in tenant '(.+?)' using json '(.+?)'$")
-    public void installServiceFromMarathonJson(String service, String model, String version, String name, String tenant, String jsonFile) throws Exception {
+    @Given("^I install service '(.+?)'( in folder '(.+?)')? with model '(.+?)' and version '(.+?)' and instance name '(.+?)' in tenant '(.+?)' using json '(.+?)'$")
+    public void installServiceFromMarathonJson(String service, String folder, String model, String version, String name, String tenant, String jsonFile) throws Exception {
         String endPoint = "/service/" + ThreadProperty.get("deploy_api_id") + "/deploy/" + service + "/" + model + "/" + version + "/schema?tenantId=" + tenant;
         String data = this.commonspec.retrieveData(jsonFile, "json");
 
@@ -985,10 +986,22 @@ public class CCTSpec extends BaseGSpec {
             endPointStatus = "/service/" + ThreadProperty.get("cct-marathon-services_id") + "/v1/services?tenant=" + tenant;
         }
 
-        String serviceName = "/" + name;
+        if (folder != null && folder.startsWith("/")) {
+            folder = folder.substring(1);
+        }
+        if (folder != null && folder.endsWith("/")) {
+            folder = folder.substring(folder.length() - 1);
+        }
 
+        String serviceName = "/" + name;
+        if (folder != null) {
+            serviceName = "/" + folder + "/" + name;
+        }
         if (!"NONE".equals(tenant)) {
             serviceName = "/" + tenant + "/" + tenant + "-" + name;
+            if (folder != null) {
+                serviceName =  "/" + tenant + "/" + folder + "/" + tenant + "-" + name;
+            }
         }
 
         restSpec.sendRequestTimeout(200, 20, "GET", endPointStatus, null, serviceName);